Semantic_ranker/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
spacy>=3.7.2,<3.8.0
python-dateutil>=2.8.2
joblib>=1.3.0
PyPDF2>=3.0.0
docx2txt>=0.8
chromadb>=0.5.3
openai>=1.30.0
langchain>=0.2.0
chainlit>=1.2.0
httpx>=0.27.0
uvicorn>=0.35.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0